pyproject.toml
[project]
name = "django-apis"
version = "0.1.0"
description = "Multiple Django Rest APIs and Projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.1",
    "attrs==25.4.0",
    "django==6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter==25.2",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.29.0",
    "gunicorn==25.1.0",
    "inflection==0.5.1",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "packaging==26.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt==2.11.0",
    "python-decouple>=3.8",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "rpds-py==0.30.0",
    "sqlparse==0.5.5",
    "uritemplate==4.2.0",
]
